+++ src/ru/yandex/practicum/delivery/model/FragileParcel.java
@@ -5,7 +5,7 @@ import ru.yandex.practicum.delivery.navigation.Address;
 import java.time.LocalDate;
 
 public class FragileParcel extends Parcel implements Trackable {
-    public FragileParcel(String description, double weight, Address deliveryAddress, LocalDate sendDay) {
+    public FragileParcel(String description, int weight, Address deliveryAddress, LocalDate sendDay) {
         super(description, weight, deliveryAddress, sendDay);
         this.deliveryType = DeliveryType.FRAGILE;
     }
